requirements.txt
Flask==2.2.2
flask_cors==4.0.0
psycopg2==2.9.6
python-dotenv==1.0.0
